Se ajustan los controladores de acuerdo al dominio del negocio de cada actor.
