--- src/main/java/co/edu/uniquindio/uniEventos/controladores/ImagenesControlador.java
@@ -1,36 +1,0 @@
-package co.edu.uniquindio.uniEventos.controladores;
-
-import co.edu.uniquindio.uniEventos.dto.MensajeDTO;
-import co.edu.uniquindio.uniEventos.servicios.interfaces.ImagenesServicio;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/imagenes")
-public class ImagenesControlador {
-
-    private final ImagenesServicio imagenesServicio;
-
-    /*
-    @RequestParam se utiliza para extraer parámetros de consulta(query
-    parameters) de la URL de una solicitud HTTP
-     */
-
-    @PostMapping("/subir")
-    public ResponseEntity<MensajeDTO<String>> subir(@RequestParam("imagen") MultipartFile imagen) throws Exception {
-        String respuesta = imagenesServicio.subirImagen(imagen);
-        return ResponseEntity.ok().body(new MensajeDTO<>(false, respuesta));
-    }
-
-    @DeleteMapping("/eliminar")
-    public ResponseEntity<MensajeDTO<String>> eliminar(@RequestParam("idImagen") String idImagen) throws Exception{
-        imagenesServicio.eliminarImagen( idImagen );
-        return ResponseEntity.ok().body(new MensajeDTO<>(false, "La imagen fue eliminada " +
-                "correctamente"));
-    }
-}
